Restructure project to remove 'app' directory nesting
