+++ filetransfer/filetransfer.go
@@ -1,6 +1,6 @@
 package filetransfer
 
-import "github.com/companieshouse/payment-reconciler/app/models"
+import "github.com/companieshouse/payment-reconciler/models"
 
 // FileTransfer provides an interface by which to upload CSV files
 type FileTransfer interface {
