+++ lambda/lambda_test.go
@@ -4,10 +4,10 @@ import (
 	"errors"
 	"testing"
 
-	"github.com/companieshouse/payment-reconciler/app/config"
-	"github.com/companieshouse/payment-reconciler/app/filetransfer"
-	"github.com/companieshouse/payment-reconciler/app/models"
-	"github.com/companieshouse/payment-reconciler/app/service"
+	"github.com/companieshouse/payment-reconciler/config"
+	"github.com/companieshouse/payment-reconciler/filetransfer"
+	"github.com/companieshouse/payment-reconciler/models"
+	"github.com/companieshouse/payment-reconciler/service"
 	"github.com/golang/mock/gomock"
 	. "github.com/smartystreets/goconvey/convey"
 )
